refactor: move locale from src/features to app/locales
