+++ app/locales/ja.ts
