+++ app/src/components/BaseMetadata.astro
@@ -1,6 +1,6 @@
 ---
 import { AUTHOR, FALLBACK_OG_IMAGE_URL, TWITTER_ACCOUNT_ID } from '../../constants/siteData';
-import { retrieveTranslation } from '../features/locales/i18n';
+import { retrieveTranslation } from '../../locales/i18n';
 
 interface Props {
   description: string;
