+++ app/src/components/ExternalResourceLink.astro
@@ -1,7 +1,7 @@
 ---
 import { pathList } from '../../constants/pathList';
 import { ASSETS_URL, FAVICON_URL } from '../../constants/siteData';
-import { retrieveTranslation } from '../features/locales/i18n';
+import { retrieveTranslation } from '../../locales/i18n';
 
 const webSiteTitle = retrieveTranslation('website.title');
 ---
