+++ app/src/data/structured_data/blogPostingStructuredData.ts
@@ -1,4 +1,4 @@
-import { retrieveTranslation } from '../../features/locales/i18n';
+import { retrieveTranslation } from '../../../locales/i18n';
 
 interface Params {
   title: string;
