+++ app/src/features/entry/components/PublicationMetadata.astro
@@ -1,5 +1,5 @@
 ---
-import { retrieveTranslation } from '../../locales/i18n';
+import { retrieveTranslation } from '../../../../locales/i18n';
 import { formatIsoString, formatYYMDString } from '../date';
 
 interface Props {
