+++ app/src/features/entry/components/Taxonomy.astro
@@ -1,6 +1,6 @@
 ---
 import { pathList } from '../../../../constants/pathList';
-import { retrieveTranslation } from '../../locales/i18n';
+import { retrieveTranslation } from '../../../../locales/i18n';
 
 interface Props {
   tags: string[];
