+++ app/src/layouts/GlobalFooter.astro
@@ -1,6 +1,6 @@
 ---
 import { pathList } from '../../constants/pathList';
-import { retrieveTranslation } from '../features/locales/i18n';
+import { retrieveTranslation } from '../../locales/i18n';
 ---
 
 <style>
