+++ app/src/layouts/GlobalHeader.astro
@@ -1,5 +1,5 @@
 ---
-import { retrieveTranslation } from '../features/locales/i18n';
+import { retrieveTranslation } from '../../locales/i18n';
 import { addWbr } from '../utils/addWbr';
 
 interface Props {
