+++ app/src/layouts/Head.astro
@@ -1,7 +1,7 @@
 ---
+import { retrieveTranslation } from '../../locales/i18n';
 import BaseMetadata from '../components/BaseMetadata.astro';
 import ExternalResourceLink from '../components/ExternalResourceLink.astro';
-import { retrieveTranslation } from '../features/locales/i18n';
 
 interface Props {
   title?: string;
