+++ app/src/pages/index.astro
@@ -2,11 +2,11 @@
 import { getCollection } from 'astro:content';
 
 import { pathList } from '../../constants/pathList';
+import { retrieveTranslation } from '../../locales/i18n';
 import EntryList from '../components/EntryList.astro';
 import { createHomeStructuredData } from '../data/structured_data/homeStructuredData';
 import { addExcerptToEntries } from '../features/entry/addExcerptToEntries';
 import { getSortedEntries } from '../features/entry/getSortedEntries';
-import { retrieveTranslation } from '../features/locales/i18n';
 import Body from '../layouts/Body.astro';
 import GlobalFooter from '../layouts/GlobalFooter.astro';
 import GlobalHeader from '../layouts/GlobalHeader.astro';
