+++ app/src/pages/policy.astro
@@ -2,8 +2,8 @@
 import '../features/entry/entryContents.css';
 
 import { pathList } from '../../constants/pathList';
+import { retrieveTranslation } from '../../locales/i18n';
 import AnalyticsOptoutSwitch from '../data/tracking/components/AnalyticsOptoutSwitch.astro';
-import { retrieveTranslation } from '../features/locales/i18n';
 import Body from '../layouts/Body.astro';
 import GlobalFooter from '../layouts/GlobalFooter.astro';
 import GlobalHeader from '../layouts/GlobalHeader.astro';
